requirements.txt
Babel==2.8.0
boto3==1.13.4
botocore==1.16.4
Django==2.2.10
django-crispy-forms==1.8.1
django-phonenumber-field==4.0.0
django-storages==1.9.1
django-multiselectfield==0.1.12
django-tempus-dominus==5.1.2.13
docutils==0.15.2
jmespath==0.9.5
mysqlclient==1.4.6
phonenumbers==8.11.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
xlwt==1.3.0
python-dotenv~=0.13.0
gunicorn==20.0.4
dj-config-url~=0.1.1
